requirements.txt
flask
flask-cors
werkzeug
python-docx
python-pptx
PyPDF2
nbformat
nbconvert
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mammoth==1.4.2
python-pptx==0.6.21
PyPDF2==3.0.1
nbformat==5.8.0
whoosh==2.7.4
